Fix chipped textures
The texture paths used by Chipped did not match any of the formats
tried by this mod. These changes introduce support for an alternate
texture path based on Chipped to resolve these issues.

It doesn't seem to entirely resolve issues with the top texture for
Chipped mushroom stems but there are no obvious missing textures when
scanning the EMI images and a proper fix may be fiddly.

The fallback handling for missing log textures has also been changed
to call the normal plank texture function, as it was calling a more
basic function that didn't try alternate paths, leading to failures.